Allen Project/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
pyttsx3==2.90
Pillow==10.2.0 